Implemented 2 layer FaaS security model
To improve security, 2 FaaS instances are used.  The first instance fetches the relevant user variables for a request and passes it to the 2nd FaaS instance.  The 2nd FaaS instance handles the actual execution of the user-provided code.  This ensures that even if the user-provided code breaks out of the sandbox, they are unable to access any sensitive data since they are in an isolated FaaS instance.  Users in the second FaaS instance also have access to saving temporary data to the tmp directory on the FaaS instance.